Cheat Menu

+++ Assets/Scripts/Library/Character/ScriptableObjects/CurrencyObject.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "Currency")]
 public class CurrencyObject : BaseScriptableObject
 {
-    public int currentCurrency;
-    public int flakLevel;
-    public int flameLevel;
-    public int mgLevel;
+    public ushort currentCurrency;
+    public ushort flakLevel;
+    public ushort flameLevel;
+    public ushort mgLevel;
 
     public bool flakActive;
     public bool flameActive;
